first

+++ project 1/python_project_1.py
@@ -1,8 +1,9 @@
+import datetime
 print("Welcome to the World of Games.")
 
 player_name=input("\nKindly enter your ingame name:\n")
 player_age=int(input("\nKindly enter your age:\n"))
-character_height=int(input("\nKindly enter  ingame Character height in meters:\n"))
+character_height=float(input("\nKindly enter  ingame Character height in meters:\n"))
 Favourite_game=input("\nPlease enter your favourite game:\n")
 Current_level=int(input("\nPlease enter your game level\n"))
 
@@ -14,7 +15,7 @@ print("\nYour Character Height is ",character_height,"[class type:",type(charact
 print("\nYour Favourite game is ",Favourite_game,"[class type:",type(Favourite_game),",Memory address:",id(Favourite_game),"]")
 print("\nYour Current level  is ",Current_level,"[class type:",type(Current_level),",Memory address:",id(Current_level),"]")
 
-import datetime
+
 current_year = datetime.datetime.now().year
 birth_year = current_year -  player_age               
 print("\nYour birth year is approximately:\n ",birth_year)
